3.1-drf-intro/smart_home/requirements.txt
django~=4.1.1
djangorestframework~=3.14.0
psycopg2-binary

python-dotenv~=0.21.0